#7 Changed default "need more" color to red
Just to make more sense and add contrast to items player should keep. Can be modified in config anyway.

Fixed an issue with fetching configs from server that would result in an exception upon ending a raid. That was caused by SessionID not being used to make the request I think. Also changed the route to /client/config/checkmark instead because /server/config/... seemed to have a special case on server that would make it return nothing. The config is still in user/configs/checkmarks.json