Fixed a bug where Session or Session.Ragfair can be null? sometimes at the start of a scav run?
really not sure where this came from, but it wasn't relevant to our code. It simply prevented us from checking if an item was in the wishlist, so if we can't check it, just assume it isn't. This rarely happens and for all I know only happened to me. There was no point in handling this error more specifically, a try is sufficient.